chore: add postgres to dagger
