+++ internal/credit/postgres_connector/balance_test.go
@@ -6,9 +6,6 @@ import (
 	"testing"
 	"time"
 
-	"entgo.io/ent/dialect"
-	entsql "entgo.io/ent/dialect/sql"
-	"github.com/peterldowns/pgtestdb"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/openmeterio/openmeter/api"
@@ -459,14 +456,7 @@ func TestPostgresConnectorBalances(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 			t.Log(tc.description)
-			driver := entsql.OpenDB(dialect.Postgres, pgtestdb.New(t, pgtestdb.Config{
-				DriverName: "pgx",
-				User:       "postgres",
-				Password:   "postgres",
-				Host:       "localhost",
-				Port:       "5432",
-				Options:    "sslmode=disable",
-			}, &EntMigrator{}))
+			driver := initDB(t)
 			databaseClient := db.NewClient(db.Driver(driver))
 			defer databaseClient.Close()
